Replaced Element by Isotope in calculation
Previously. we used Elements in the calculation because in the final
results only Elements should be displayed. This led, however, to the
problem that in the check whether a combination of elements was already
contained in the result or backlog, the Calculator would not add newly
found molecules even if the isotopes were different. This way, we lost
possible results.

I replaced the elements by isotopes and adapated the rest of the program
accordingly. The test run correctly.